+++ Watchman.Discord/Areas/Help/Services/HelpDBGeneratorService.cs
@@ -1,79 +1,25 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using Devscord.DiscordFramework.Middlewares.Contexts;
 using Devscord.DiscordFramework.Services;
 using Devscord.DiscordFramework.Services.Models;
-using Newtonsoft.Json;
-using Watchman.Cqrs;
 using Watchman.Discord.Areas.Help.Factories;
 using Watchman.DomainModel.Help.Models;
-using Watchman.DomainModel.Help.Queries;
 using Watchman.Integrations.MongoDB;
 
 namespace Watchman.Discord.Areas.Help.Services
 {
-    public class HelpService : IService
+    public class HelpDBGeneratorService : IService
     {
-        private readonly IQueryBus _queryBus;
         private readonly HelpInformationFactory _helpInformationFactory;
         private readonly ISession _session;
 
-        public HelpService(ISessionFactory sessionFactory, IQueryBus queryBus, HelpInformationFactory helpInformationFactory)
+        public HelpDBGeneratorService(ISessionFactory sessionFactory, HelpInformationFactory helpInformationFactory)
         {
-            _queryBus = queryBus;
             _helpInformationFactory = helpInformationFactory;
             _session = sessionFactory.Create();
         }
 
-        //todo: tutaj myślę czy nie wydzielić generateHelp i generateJsonHelp do osobnej klasy
-        public string GenerateHelp(Contexts contexts)
-        {
-            var result = this._queryBus.Execute(new GetHelpInformationQuery(contexts.Server.Id));
-
-            var messageBuilder = new StringBuilder();
-            foreach (var helpInfo in result.HelpInformations)
-            {
-                var line = new StringBuilder("-" + helpInfo.Names.Aggregate((x, y) => $"{x} / -{y}"));
-                line.Append(" => ");
-                line.Append(helpInfo.Descriptions.First(x => x.Name == helpInfo.DefaultDescriptionName).Details);
-                messageBuilder.AppendLine(line.ToString());
-            }
-
-            return messageBuilder.ToString();
-        }
-
-        public IEnumerable<string> GenerateJsonHelp(Contexts contexts)
-        {
-            var result = this._queryBus.Execute(new GetHelpInformationQuery(contexts.Server.Id));
-            var serialized = JsonConvert.SerializeObject(result.HelpInformations, Formatting.Indented);
-            var smallerMessages = SplitJsonToSmallerMessages(serialized);
-
-            foreach (var sm in smallerMessages)
-            {
-                var messageBuilder = new StringBuilder();
-                messageBuilder.AppendLine("```json");
-                messageBuilder.AppendLine(sm);
-                messageBuilder.AppendLine("```");
-                yield return messageBuilder.ToString();
-            }
-        }
-
-        private IEnumerable<string> SplitJsonToSmallerMessages(string message)
-        {
-            const int MAX_MESSAGE_LENGTH = 1998; // for safety reason I made it smaller
-
-            while (message.Length > MAX_MESSAGE_LENGTH)
-            {
-                var cutIndex = message.Substring(0, MAX_MESSAGE_LENGTH).LastIndexOf("},", StringComparison.Ordinal) + 2;
-                yield return message.Substring(0, cutIndex);
-                message = message.Remove(0, cutIndex);
-            }
-            yield return message;
-        }
-
-        public void FillDatabaseWithNewMethods(IEnumerable<CommandInfo> commandInfosFromAssembly)
+        public void SyncDatabaseWithCode(IEnumerable<CommandInfo> commandInfosFromAssembly)
         {
             var commandInfosFromAssemblyList = commandInfosFromAssembly.ToList(); // for not multiple enumerating
             var helpInfosInDb = _session.Get<HelpInformation>().ToList();
